src/requirements.txt
gradio==4.24.0
boto3==1.34.72
openai==1.14.3
python-dotenv==1.0.1